i hate this, end my suffering

+++ Assets/Scripts/Inventory/InventorySlot.cs
@@ -5,22 +5,22 @@ public class InventorySlot : MonoBehaviour
 {
     [field: SerializeField] public bool IsOccupied { get; private set; }
     [field: SerializeField] public ItemSO ItemInSlot { get; private set; }
-    [field: SerializeField] public ushort ItemAmount { get; private set; }
+    [field: SerializeField] public int ItemAmount { get; private set; }
 
     private void changeOccupancy(bool occupancyValue)
     {
         IsOccupied = occupancyValue;
     }
-    public void AddItemAmount(ushort amount)
+    public void AddItemAmount(int amount)
     {
         ItemAmount += amount;
     }
-    public void RemoveItemAmount(ushort amount)
+    public void RemoveItemAmount(int amount)
     {
         ItemInSlot = null;
         ItemAmount -= amount;
     }
-    public void AddItem(ItemSO item, ushort amount)
+    public void AddItem(ItemSO item, int amount)
     {
         ItemInSlot = item;
         ItemAmount += amount;
